reorganizing everything, just need to add definitions
